+++ tests/testthat/test-replace_nan_inf.R
@@ -24,4 +24,16 @@ test_that("extract from data frame", {
       class = "data.frame"
     )
   )
+
+  expect_equal(
+    replace_nan_inf(df, select = starts_with("x")),
+    structure(
+      list(
+        x = c(1, NA, 5, NA, 2, NA),
+        y = c(3, NaN, 4, -Inf, 6, 7)
+      ),
+      row.names = c(NA, -6L),
+      class = "data.frame"
+    )
+  )
 })